tag packages as LRL
In https://github.com/RichardLitt/endangered-languages/issues/63 we discovered that it was nearly impossible to find packages which pertained to low resource languages. 

We suggest that if you have a project that could be used for other low resource languages you use the tag/keyword "lrl" so that your package is easy to search for and find.